Refactor delete command

+++ commands/hubdb/delete.ts
@@ -4,27 +4,49 @@ const { logError } = require('../../lib/errorHandlers/index');
 const { deleteTable } = require('@hubspot/local-dev-lib/api/hubdb');
 const { loadAndValidateOptions } = require('../../lib/validation');
 const { trackCommandUsage } = require('../../lib/usageTracking');
-
+const { addUseEnvironmentOptions } = require('../../lib/commonOpts');
 const {
-  addConfigOptions,
-  addAccountOptions,
-  addUseEnvironmentOptions,
-} = require('../../lib/commonOpts');
+  selectHubDBTablePrompt,
+} = require('../../lib/prompts/selectHubDBTablePrompt');
+const { promptUser } = require('../../lib/prompts/promptUtils');
+const { EXIT_CODES } = require('../../lib/enums/exitCodes');
 const { i18n } = require('../../lib/lang');
 
 const i18nKey = 'commands.hubdb.subcommands.delete';
 
-exports.command = 'delete <tableId>';
+exports.command = 'delete [table-id]';
 exports.describe = i18n(`${i18nKey}.describe`);
 
 exports.handler = async options => {
-  const { tableId, derivedAccountId } = options;
+  const { force, derivedAccountId } = options;
 
   await loadAndValidateOptions(options);
 
   trackCommandUsage('hubdb-delete', null, derivedAccountId);
 
   try {
+    const { tableId } =
+      'tableId' in options
+        ? options
+        : await selectHubDBTablePrompt({
+            accountId: derivedAccountId,
+            options,
+            isDeleteCommand: true,
+          });
+
+    if (!force) {
+      const { shouldDeleteTable } = await promptUser({
+        name: 'shouldDeleteTable',
+        type: 'confirm',
+        message: i18n(`${i18nKey}.shouldDeleteTable`, { tableId }),
+      });
+      process.stdin.resume();
+
+      if (!shouldDeleteTable) {
+        process.exit(EXIT_CODES.SUCCESS);
+      }
+    }
+
     await deleteTable(derivedAccountId, tableId);
     logger.success(
       i18n(`${i18nKey}.success.delete`, {
@@ -32,6 +54,7 @@ exports.handler = async options => {
         tableId,
       })
     );
+    process.exit(EXIT_CODES.SUCCESS);
   } catch (e) {
     logger.error(
       i18n(`${i18nKey}.errors.delete`, {
@@ -43,12 +66,15 @@ exports.handler = async options => {
 };
 
 exports.builder = yargs => {
-  addAccountOptions(yargs);
-  addConfigOptions(yargs);
   addUseEnvironmentOptions(yargs);
 
-  yargs.positional('tableId', {
+  yargs.positional('table-id', {
     describe: i18n(`${i18nKey}.positionals.tableId.describe`),
     type: 'string',
+  });
+
+  yargs.option('force', {
+    describe: i18n(`${i18nKey}.options.force.describe`),
+    type: 'boolean',
   });
 };
